Orderbook updates saved locally

+++ include/trading/Order.h
@@ -4,5 +4,5 @@ struct Order {
     double price;
     long volume;
 
-    Order(double p, long v);
+    Order(double p, double v);
 };
